Moving getLevelIncreases into Item
